+++ home/src/environments/environment.prod.ts
@@ -26,7 +26,7 @@ export const environment = {
       xeerpa: '/v1/xeerpa'
     }
   },
-  gtmId: 'GTM-MGL9B93',
+  googleTagManagerId: 'GTM-MGL9B93',
   googleAnalyticsId: 'UA-159252784-5',
   hosting: 'https://tapit-sso-production.web.app/v3',
   facebook: {