+++ home/src/environments/environment.ts
@@ -30,7 +30,7 @@ export const environment = {
       xeerpa: '/v1/xeerpa'
     }
   },
-  gtmId: 'GTM-KQF533X',
+  googleTagManagerId: 'GTM-KQF533X',
   googleAnalyticsId: 'UA-159252784-6',
   hosting: '',
   facebook: {